Only pull and push in the right network

+++ .hammerspoon/sleep-hooks.lua
@@ -2,18 +2,16 @@ require('utils')
 
 
 function pre_lock()
-    hs.execute("killall newsboat", true)
-    hs.execute("make -f $HOME/Makefile push", true)
-    
+    if (hs.wifi.currentNetwork() == "Wachtberg") then
+        hs.execute("killall newsboat", true)
+        hs.execute("make -f $HOME/Makefile push", true)
+    end
+
     device = hs.audiodevice.defaultOutputDevice()
     device:setOutputVolume(5)
     hlspeak('doop')
 end
 
-function do_pull()
-    hs.execute("make -f $HOME/Makefile pull", true)
-    hlspeak('hello')
-end
 
 function post_lock()
     device = hs.audiodevice.defaultOutputDevice()
@@ -22,10 +20,9 @@ function post_lock()
     hlspeak('bell')
 
     print("Current WIFI is "..hs.wifi.currentNetwork())
-    if (not hs.wifi.currentNetwork()) then
-        hs.timer.doAfter(3, do_pull)
-    else
-        do_pull()
+    if (hs.wifi.currentNetwork() == "Wachtberg") then
+        hs.execute("make -f $HOME/Makefile pull", true)
+        hlspeak('hello')
     end
 end
 
@@ -43,7 +40,6 @@ end
 
 sleepWatcher = hs.caffeinate.watcher.new(sleepWatch)
 sleepWatcher:start()
-
